(PR#39793) RFE: civ3-like land support
Add more extensive land terrain for editor.  It attempts to harmonize
the existing civ1-like and civ2-like rulesets, that inexplicably use
different identifiers in savegames.  Mismatched identifiers are
automatically converted.

For references in terrain.ruleset files, use the full section label, 
instead of the name field.  This allows multiple entries with the 
same descriptive name, to support rulesets that need to treat them 
the same, or want to use fantasy or fictional names.  Also, supports a 
more generic table driven nation city_name system.

Sorted the resources in the same fashion as the terrain (by identifier) 
to ease future terrain.ruleset comparisons.

Note that the use of graphics_alt means that none of the new terrain
will actually show up, until there are more graphics.  But this has to
commit first, so that artists can work on the graphics....


git-svn-id: http://svn.gna.org/svn/freeciv/trunk@13895 a0f10bec-cc02-0410-94fc-a9cfff90b4cd