(PR#40116) client CMA crash for trade routes during game load
reported by:
  Christian Knoke <chrisk@cknoke.de>

The proximate issue is that the CMA is trying to calculate a trade route
value to another city, but that city hasn't been sent to the client yet.
This rare ordering bug is easily fixed by checking the city tile.

Also, the call to the agent was before city_info_common() was called; so
the city itself isn't in the player->cities list, other initialization
isn't done, etc.  Fixed.

The ultimate issue is that the agent code is trying to calculate values
during load before everything knowable is known.  This results in
different values than during game play!  Not fixed!


git-svn-id: http://svn.gna.org/svn/freeciv/trunk@14452 a0f10bec-cc02-0410-94fc-a9cfff90b4cd