Center map view at start of first game (GTK+ client)
The Gtk client did have the startunits in the corner of the map view
when starting the first game of the same client process. This was due
to the client centering the view before the canvas widget was given
its size.

Now we call center_on_something on the canvas' first configure-event.

Remove the map_configure variable, whose value is never used.

Bug reported by Matthias Pfafferodt <syntron>

See gna bug #15738

git-svn-id: http://svn.gna.org/svn/freeciv/trunk@17211 a0f10bec-cc02-0410-94fc-a9cfff90b4cd