Remove unnecessary padding around unit_info_label
This is the second patch in a short series intended to improve the gtk
gui on netbook sized displays.

In the gtk gui the unit_info_label is wrapped in a scrolling window that
never scrolls, and thus only adds ugly and unnecessary padding around
the label.

This patch removes the superfluous scrolling window.

patch by Jon Severinsson (jon_severinsson)

see gna patch #1685



git-svn-id: http://svn.gna.org/svn/freeciv/trunk@17474 a0f10bec-cc02-0410-94fc-a9cfff90b4cd