new setting plrcolormode + command /playercolor
* How to pick the player color (plrcolormode):
 - "player color (ordered)" (PLR_ORDER): select the color for each
   player according to the order of the color definition.
 - "player color (random)" (PLR_RANDOM): select a random color for each
   player.
 - "player color (set/random)" (PLR_SET): use the color set via the
   playercolor command. For players without a color a random value will
   be selected.
 - "team color (ordered)" (TEAM_ORDER): select the color for one team
   depending on the order of the color definition.

* add argument color to list command

* command /playercolor:

If a player name is given only the color for this palyer is shown. The
player color can be set if a color definition is added. The color
displayed by the client will change in the next turn. A color is defined
as a hex string, i.e '#ff0000' for red.

see gna patch #2062


git-svn-id: http://svn.gna.org/svn/freeciv/trunk@19151 a0f10bec-cc02-0410-94fc-a9cfff90b4cd