[unused var fix] macros in map.h
the macros define in map.h iterate also over the tile; thus, this
variable is set but sometimes not used. This patch adds reduced macros
for this case. Perhaps the macro definitions of 'adjc_dirlist_iterate'
and 'adjc_dirlist_base_iterate' can be merged? I did not find a
solution.

see gna bug #19152



git-svn-id: http://svn.gna.org/svn/freeciv/trunk@20691 a0f10bec-cc02-0410-94fc-a9cfff90b4cd